判別さんのクラス名を変更
* コインメックという自販機の機構の名称を使った。
* 綴りが合っているか微妙だったが。
